Most functions working
Pot value is being read and displayed.
Switched to using two screens for each value: One for the label, one for the value. Gave up on using a prefix.

Need to add code to make the pressure transmitter re-zero, and make it display the delta. Need to also test with an actual fan.